# openapi.yaml
openapi: 3.0.3
info:
  title: GoGuia Auth API
  version: 1.0.0
  description: |
    Autenticação usada pelo frontend Next.js.
    Base local: http://localhost:3001
servers:
  - url: http://localhost:3001
    description: Desenvolvimento
tags:
  - name: Auth
  - name: Health
paths:
  /health:
    get:
      tags: [Health]
      summary: Saúde da API
      responses:
        '200':
          description: API no ar
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /auth/cadastro:
    post:
      tags: [Auth]
      summary: Cadastro
      description: Aceita JSON sem foto ou multipart/form-data com o campo opcional `foto`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cadastro'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CadastroMultipart'
      responses:
        '201':
          description: Usuário criado. Com REQUIRE_EMAIL_CONFIRMATION=false, inclui token.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/AuthSuccess'
                  - $ref: '#/components/schemas/CadastroPendente'
        '400':
          $ref: '#/components/responses/Validation'
        '409':
          description: E-mail já cadastrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: E-mail já cadastrado
                code: EMAIL_ALREADY_EXISTS
                errors:
                  email: E-mail já cadastrado
        '413':
          description: Foto acima de 5MB
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Arquivo excede o limite
                code: FILE_TOO_LARGE
        '415':
          description: Formato de imagem não suportado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Formato de imagem não suportado
                code: UNSUPPORTED_MEDIA_TYPE
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
      responses:
        '200':
          description: Sessão criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccess'
        '400':
          $ref: '#/components/responses/Validation'
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Credenciais inválidas
                code: INVALID_CREDENTIALS
        '403':
          description: E-mail não confirmado, só quando REQUIRE_EMAIL_CONFIRMATION=true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Confirme seu e-mail antes de entrar
                code: EMAIL_NOT_CONFIRMED
  /auth/me:
    get:
      tags: [Auth]
      summary: Usuário da sessão
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Usuário autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout stateless
      description: Confirma o Bearer e responde 204. O JWT continua válido até expirar; o frontend deve apagar o token.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Logout aceito
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/esqueci-senha:
    post:
      tags: [Auth]
      summary: Pedido de redefinição de senha
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Sempre a mesma mensagem quando o e-mail tem formato válido
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Se o e-mail existir, enviaremos instruções para redefinir a senha
        '400':
          $ref: '#/components/responses/Validation'
  /auth/redefinir-senha:
    post:
      tags: [Auth]
      summary: Redefine a senha com o token do e-mail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, novaSenha]
              properties:
                token:
                  type: string
                novaSenha:
                  type: string
      responses:
        '200':
          description: Senha alterada
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Senha alterada com sucesso
        '400':
          description: Token inválido ou senha fraca
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/confirmar-email:
    post:
      tags: [Auth]
      summary: Confirma o e-mail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        '200':
          description: E-mail confirmado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Confirmacao'
        '400':
          description: Token inválido ou expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags: [Auth]
      summary: Confirma o e-mail pelo link
      description: Com Accept text/html, redireciona para o frontend. Nos demais casos, devolve JSON.
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: E-mail confirmado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Confirmacao'
        '302':
          description: Redirect para /confirmacao-email no frontend
        '400':
          description: Token inválido ou expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Validation:
      description: Dados inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Dados inválidos
            code: VALIDATION_ERROR
            errors:
              email: E-mail inválido
    Unauthorized:
      description: Token ausente, inválido ou expirado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Token ausente, inválido ou expirado
            code: UNAUTHORIZED
  schemas:
    User:
      type: object
      required: [id, nomeCompleto, email, fotoUrl, tipo, emailConfirmado]
      properties:
        id:
          type: string
          format: uuid
        nomeCompleto:
          type: string
        email:
          type: string
          format: email
        fotoUrl:
          type: string
          nullable: true
        tipo:
          type: string
          enum: [TURISTA, GUIA]
        emailConfirmado:
          type: boolean
    AuthSuccess:
      type: object
      required: [token, expiresIn, user]
      properties:
        token:
          type: string
        expiresIn:
          type: integer
          example: 86400
        user:
          $ref: '#/components/schemas/User'
    CadastroPendente:
      type: object
      required: [message, user]
      properties:
        message:
          type: string
        user:
          $ref: '#/components/schemas/User'
    Confirmacao:
      type: object
      properties:
        message:
          type: string
          example: E-mail confirmado
        user:
          $ref: '#/components/schemas/User'
    Cadastro:
      type: object
      required: [nomeCompleto, email, senha]
      properties:
        nomeCompleto:
          type: string
          minLength: 2
          maxLength: 120
        email:
          type: string
          format: email
        senha:
          type: string
          format: password
          minLength: 8
        tipo:
          type: string
          enum: [TURISTA, GUIA]
          default: TURISTA
    CadastroMultipart:
      allOf:
        - $ref: '#/components/schemas/Cadastro'
        - type: object
          properties:
            foto:
              type: string
              format: binary
              description: JPG, PNG ou WEBP, até 5MB
    Login:
      type: object
      required: [email, senha]
      properties:
        email:
          type: string
          format: email
        senha:
          type: string
          format: password
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string